# backend-auth/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Authentication and Authorization microservice for bus booking system",
        "title": "Bus Booking Auth Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@busbooking.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/reviews/my-reviews": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all reviews by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Get user reviews",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reviews/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing review (owner only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Update a review",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Review ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated review data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an existing review (owner only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Delete a review",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Review ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/trips/{id}/reviews": {
            "get": {
                "description": "Get paginated reviews for a specific trip",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Get trip reviews",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 10, max 50)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecases.TripReviewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a review for a completed trip (authenticated users only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Create a trip review",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Review data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.Booking": {
            "type": "object",
            "properties": {
                "booking_reference": {
                    "description": "Unique booking reference (e.g., \"BK20231215ABC123\")",
                    "type": "string"
                },
                "cancellation_reason": {
                    "description": "Reason for cancellation",
                    "type": "string"
                },
                "cancelled_at": {
                    "description": "When booking was cancelled",
                    "type": "string"
                },
                "confirmed_at": {
                    "description": "When booking was confirmed",
                    "type": "string"
                },
                "contact_email": {
                    "description": "Primary contact email",
                    "type": "string"
                },
                "contact_name": {
                    "description": "Primary contact name",
                    "type": "string"
                },
                "contact_phone": {
                    "description": "Primary contact phone",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "expires_at": {
                    "description": "Expiration time for pending bookings",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_guest_booking": {
                    "description": "True if booked without account",
                    "type": "boolean"
                },
                "notes": {
                    "description": "Additional notes",
                    "type": "string"
                },
                "passengers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Passenger"
                    }
                },
                "payment_method": {
                    "description": "e.g., \"credit_card\", \"cash\", \"bank_transfer\"",
                    "type": "string"
                },
                "payment_reference": {
                    "description": "External payment reference",
                    "type": "string"
                },
                "payment_status": {
                    "$ref": "#/definitions/entities.PaymentStatus"
                },
                "status": {
                    "$ref": "#/definitions/entities.BookingStatus"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Ticket"
                    }
                },
                "total_amount": {
                    "description": "Total price",
                    "type": "number"
                },
                "total_seats": {
                    "description": "Number of seats booked",
                    "type": "integer"
                },
                "trip": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.Trip"
                        }
                    ]
                },
                "trip_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/entities.User"
                },
                "user_id": {
                    "description": "Nullable for guest bookings",
                    "type": "string"
                }
            }
        },
        "entities.BookingStatus": {
            "type": "string",
            "enum": [
                "pending",
                "confirmed",
                "cancelled",
                "completed",
                "expired"
            ],
            "x-enum-comments": {
                "BookingStatusCancelled": "Booking cancelled",
                "BookingStatusCompleted": "Trip completed",
                "BookingStatusConfirmed": "Payment confirmed",
                "BookingStatusExpired": "Booking expired due to non-payment",
                "BookingStatusPending": "Booking created but not paid"
            },
            "x-enum-descriptions": [
                "Booking created but not paid",
                "Payment confirmed",
                "Booking cancelled",
                "Trip completed",
                "Booking expired due to non-payment"
            ],
            "x-enum-varnames": [
                "BookingStatusPending",
                "BookingStatusConfirmed",
                "BookingStatusCancelled",
                "BookingStatusCompleted",
                "BookingStatusExpired"
            ]
        },
...
        "entities.Passenger": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "Passenger age",
                    "type": "integer"
                },
                "booking_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "description": "Passenger email",
                    "type": "string"
                },
                "full_name": {
                    "description": "Passenger full name",
                    "type": "string"
                },
                "gender": {
                    "description": "Male/Female/Other",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "id_number": {
                    "description": "ID/Passport number",
                    "type": "string"
                },
                "phone": {
                    "description": "Passenger phone",
                    "type": "string"
                },
                "seat": {
                    "$ref": "#/definitions/entities.Seat"
                },
                "seat_id": {
                    "description": "Which seat on the seat map",
                    "type": "string"
                },
                "seat_number": {
                    "description": "Denormalized for convenience",
                    "type": "string"
                },
                "seat_price": {
                    "description": "Price for this specific seat",
                    "type": "number"
                },
                "seat_type": {
                    "$ref": "#/definitions/entities.SeatType"
                },
                "special_needs": {
                    "description": "Special requirements",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "completed",
                "failed",
                "refunded"
            ],
            "x-enum-varnames": [
                "PaymentStatusPending",
                "PaymentStatusCompleted",
                "PaymentStatusFailed",
                "PaymentStatusRefunded"
            ]
        },
        "entities.Review": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/entities.Booking"
                },
                "booking_id": {
                    "description": "Verify user completed the trip",
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "rating": {
                    "description": "1-5 stars",
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "trip": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.Trip"
                        }
                    ]
                },
                "trip_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/entities.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entities.ReviewStats": {
            "type": "object",
            "properties": {
                "average_rating": {
                    "type": "number"
                },
                "rating_1_count": {
                    "type": "integer"
                },
                "rating_2_count": {
                    "type": "integer"
                },
                "rating_3_count": {
                    "type": "integer"
                },
                "rating_4_count": {
                    "type": "integer"
                },
                "rating_5_count": {
                    "type": "integer"
                },
                "total_reviews": {
                    "type": "integer"
                }
            }
        },
        "entities.Role": {
            "type": "string",
            "enum": [
                "guest",
                "passenger",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleGuest",
                "RolePassenger",
                "RoleAdmin"
            ]
        },
...
        "entities.Trip": {
            "type": "object",
            "properties": {
                "bus": {
                    "$ref": "#/definitions/entities.Bus"
                },
                "bus_id": {
                    "description": "Foreign key to buses (nullable before assignment)",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "driver_id": {
                    "description": "Foreign key to driver (future enhancement)",
                    "type": "string"
                },
                "end_time": {
                    "description": "Expected arrival date and time",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "description": "Admin notes",
                    "type": "string"
                },
                "price": {
                    "description": "Price for this specific trip",
                    "type": "number"
                },
                "route": {
                    "description": "Relations (will be loaded separately)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.Route"
                        }
                    ]
                },
                "route_id": {
                    "description": "Foreign key to routes",
                    "type": "string"
                },
                "start_time": {
                    "description": "Departure date and time",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entities.TripStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "entities.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "oauth_id": {
                    "type": "string"
                },
                "oauth_provider": {
                    "description": "google, github",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/entities.Role"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateReviewRequest": {
            "type": "object",
            "required": [
                "booking_id",
                "rating"
            ],
            "properties": {
                "booking_id": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.UpdateReviewRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "usecases.TripReviewsResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Review"
                    }
                },
                "stats": {
                    "$ref": "#/definitions/entities.ReviewStats"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}